+++ types/wali/builtin/dir.d.lua
@@ -1,0 +1,7 @@
+---@meta
+---@module 'wali.builtin.dir'
+
+---@type WaliModule<WaliBuiltinDirArgs>
+local mod = ...
+
+return mod