+++ types/wali/builtin/permissions.d.lua
@@ -1,0 +1,7 @@
+---@meta
+---@module 'wali.builtin.permissions'
+
+---@type WaliModule<WaliBuiltinPermissionsArgs>
+local mod = ...
+
+return mod