+++ types/wali/builtin/push_file.d.lua
@@ -1,0 +1,7 @@
+---@meta
+---@module 'wali.builtin.push_file'
+
+---@type WaliModule<WaliBuiltinPushFileArgs>
+local mod = ...
+
+return mod