+++ types/wali/builtin/remove.d.lua
@@ -1,0 +1,7 @@
+---@meta
+---@module 'wali.builtin.remove'
+
+---@type WaliModule<WaliBuiltinRemoveArgs>
+local mod = ...
+
+return mod