+++ types/wali/builtin/template.d.lua
@@ -1,0 +1,7 @@
+---@meta
+---@module 'wali.builtin.template'
+
+---@type WaliModule<WaliBuiltinTemplateArgs>
+local mod = ...
+
+return mod